Removed the search bar from Trending Companies & renamed to Trade Zone
