+++ src/components/TradeZone/StockGraph.vue
